search UI update
