+++ src/components/Routing.tsx
@@ -1,12 +1,12 @@
 import { Route, Routes } from 'react-router-dom';
 import LoginForm from './LoginForm';
-import Search from './SearchPage';
+import { HomePage, SearchBar } from './SearchPage';
 const Routing = () =>{
     return(
         <div>
             <Routes>
-                <Route path='/' element={ <Search.SearchBar />}/>
-                <Route path='/@:id' element={<Search.SearchPage  />}/>
+                <Route path='/' element={ <SearchBar />}/>
+                <Route path='/@:id' element={<HomePage  />}/>
                 <Route path='/login' element={<LoginForm/>}/>
             </Routes>
         </div>
