+++ src/components/cards/interfaces.tsx
@@ -1,0 +1,25 @@
+import * as Hive from '@hiveio/dhive';
+
+export interface IAddAccountKeyProps {
+    authAccountType: string;
+    setNewAccount: Function;
+  }
+  
+export interface IAccountKeysCardProps {
+authorityName: string;
+authAccountType: string;
+accountKeyAuths: [string, number][];
+}
+
+export interface IAccountKeyRowProps {
+authorityName: string;
+type: string;
+accountKeyAuth?: [string, number];
+threshold?: number;
+isLoggedIn?:boolean;
+}
+
+export interface IAuthorityCardProps {
+authorityName: string;
+authority: Hive.AuthorityType;
+}
