+++ src/components/modals/UpdateAuthoritiesConfirmation.tsx
@@ -1,4 +1,3 @@
-import * as Hive from '@hiveio/dhive';
 import { useEffect, useState } from 'react';
 import { Form, InputGroup } from 'react-bootstrap';
 import Button from 'react-bootstrap/Button';
@@ -22,61 +21,35 @@ function UpdateAuthoritiesConfirmation({show, handleClose}:Iprops) {
     const isOwnerAuthUpdated = useAppSelector(
       (state) => state.updateAuthorities.isOwnerAuthUpdated,
     );
+    const isPostingAuthUpdated =useAppSelector(
+      (state) => state.updateAuthorities.isPostingAuthUpdated,
+    );
     const[origAuths, setOrigAuths] = useState<Authorities>(originalAuthorities)
     const[newAuths, setNewAuths] = useState<Authorities>(newAuthorities)
-    const[key, setKey] =  useState<string>('');
     const[isOwnerUpdate, setIsOwnerUpdated] = useState<boolean>(isOwnerAuthUpdated);
-    const[ownerAccountsDiff, setOwnerAccountsDiff] = useState<[string|Hive.PublicKey,number][]>([...origAuths.owner.account_auths]);
-    const[ownerKeysDiff, setOwnerKeysDiff] = useState<[string|Hive.PublicKey,number][]>([...origAuths.owner.key_auths]);
-    const[activeAccountsDiff, setActiveAccountsDiff] = useState<[string|Hive.PublicKey,number][]>([...origAuths.active.account_auths]);
-    const[activeKeysDiff, setActiveKeysDiff] = useState<[string|Hive.PublicKey,number][]>([...origAuths.active.key_auths]);
-    const[postingAccountsDiff, setPostingAccountsDiff] = useState<[string|Hive.PublicKey,number][]>([...origAuths.posting.account_auths]);
-    const[postingKeysDiff, setPostingKeysDiff] = useState<[string|Hive.PublicKey,number][]>([...origAuths.posting.account_auths]);
-
-    const update = (origAuth:[string|Hive.PublicKey,number][], newAuth:[string|Hive.PublicKey,number][], setter:Function) => {
-        setter((prev:[string|Hive.PublicKey,number][]) => {
-            const newAuths = [...prev];
-            for(var i = 0; i< newAuth.length; i++){
-                if(origAuth[i][1] !== newAuth[i][1]){
-                    newAuths[i]=[...newAuth[i]];
-                    newAuths[i][1] = newAuths[i][1];
-                }
-            }
-            return newAuths;
-        })
-    }
+    const[isPostingUpdate, setIsPostingUpdate] = useState<boolean>(isPostingAuthUpdated);
+    const[key, setKey] =  useState<string>('');
 
     useEffect(()=>{
-      console.log(key);
       dispatch(setOwnerKey(key));
     },[key])
 
-    useEffect(()=>{
-        setOrigAuths({...originalAuthorities});
-    },[originalAuthorities])
-
-    useEffect(()=>{
-        setNewAuths({...newAuthorities});
-        update(origAuths.owner.account_auths, newAuths.owner.account_auths, setOwnerAccountsDiff);
-        update(origAuths.owner.key_auths, newAuths.owner.key_auths, setOwnerKeysDiff);
-        update(origAuths.active.account_auths, newAuths.active.account_auths, setActiveAccountsDiff);
-        update(origAuths.active.key_auths, newAuths.active.key_auths, setActiveKeysDiff);
-        update(origAuths.posting.account_auths, newAuths.posting.account_auths, setPostingAccountsDiff);
-        update(origAuths.posting.key_auths, newAuths.posting.key_auths, setPostingKeysDiff);
-        console.log("newAuthorities.posting: ",newAuthorities.posting)
-    },[newAuthorities])
-   
     useEffect(() => {
       setIsOwnerUpdated(isOwnerAuthUpdated);
     },[isOwnerAuthUpdated])
-    
-    useEffect(() => {
-        if(show)
-        {
-           
-        }
-    },[show])
-    
+    useEffect(()=>{
+      setIsPostingUpdate(isPostingAuthUpdated)
+    },[isPostingUpdate])
+
+
+    const handleUpdate = () =>{
+      if(isOwnerUpdate){
+        //usedhive
+      }else{
+        //usekeychain
+      }
+    }
+
     return (
     <div
       className="modal show"
@@ -92,7 +65,7 @@ function UpdateAuthoritiesConfirmation({show, handleClose}:Iprops) {
           <Modal.Title>Update Account Authorities</Modal.Title>
         </Modal.Header>
         <Modal.Body>
-          {isOwnerAuthUpdated?
+          {isOwnerUpdate?
             <OnwerKeyInput setOwnerKey={setKey}/>
           :  <div>Are you sure you wanna update?</div>
         }
@@ -101,7 +74,7 @@ function UpdateAuthoritiesConfirmation({show, handleClose}:Iprops) {
           <Button variant="secondary" onClick={()=> {handleClose()}}>
             Close
           </Button>
-          <Button variant="primary">Update</Button>
+          <Button variant="primary" onClick={()=>{handleUpdate()}}>Update</Button>
         </Modal.Footer>
       </Modal>
     </div>
